+++ Commercially/Vinconomy/Network/Packets/LedgerEntryRequestPacket.cs
@@ -7,7 +7,7 @@ namespace Vinconomy.Network.Packets
     {
 
         public LedgerEntryRequestPacket() { }
-        public LedgerEntryRequestPacket(int shopId, int month, int year)
+        public LedgerEntryRequestPacket(long shopId, int month, int year)
         {
             ShopId = shopId;
             Month = month;
@@ -15,7 +15,7 @@ namespace Vinconomy.Network.Packets
         }
 
         [ProtoMember(1)]
-        public int ShopId { get; set; }
+        public long ShopId { get; set; }
         [ProtoMember(2)]
         public int Month { get; set; }
         [ProtoMember(3)]
